Use same audit command for resolution (#1)
In commit 71d58a2 we updated the actual audit command to add --groups dependencies so that it only audits production dependencies. We didn't update the resolution instructions, so in commonlit/commonlit#24998 when I tried to add vulnerabilities to the ignore list I ended up with a much larger diff than I expected.

This PR updates the copy-pastable instructions so that the results align closer to what you would expect from reading the output of the action.